requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1

# Data Processing & Analytics
pandas==2.1.3
polars==0.19.19
numpy==1.25.2
duckdb==0.9.2
pyarrow==14.0.1

# Database & Storage
psycopg2-binary==2.9.9
sqlalchemy>=1.4.28,<2.0
alembic==1.12.1

# dbt for ELT Pipeline (replaces Airflow)
dbt-core==1.7.4
dbt-postgres==1.7.4

# ETL & HTTP
requests==2.31.0
python-dotenv==1.0.0
httpx==0.25.2
aiohttp==3.9.1

# Data Visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Machine Learning & Analysis
scikit-learn==1.3.2
scipy==1.11.4

# Utilities
pydantic>=1.10.0,<3.0.0
python-dateutil==2.8.2
schedule==1.2.0
click==8.1.7

# Development & Testing
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Jupyter for analysis
jupyter==1.0.0
ipykernel==6.26.0

# Authentication
supabase>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
email-validator>=2.0.0 